Move image format parsing to its own file
Also add a few extra checks for malformed data.
